Move resource creation for test to a fixture, add to test_app
